% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqua_light.R
\name{aqua_light}
\alias{aqua_light}
\title{Model to predict photosynthetically active radiation at the benthic surface.}
\usage{
aqua_light(
  driver_file,
  Lat,
  Lon,
  channel_azimuth,
  bankfull_width,
  BH,
  TH,
  overhang,
  overhang_height,
  x_LAD
)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{channel_azimuth}{Channel azimuth}

\item{bankfull_width}{Bankfull width}

\item{BH}{Bank height}

\item{TH}{Tree height}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}

\item{x_LAD}{Leaf angle distribution, default = 1}

\item{WL}{Water level}
}
\value{
Returns a time series of predicted photosynthetically active radiation
}
\description{
This function builds on the stream_light function by adding in
several key components. First, it makes some modifications to allow better
handling of dynamic wetted widths. Secondly, it moves beyond making predictions
at the stream surface and includes the influence of surface reflection and
attenuation as a function of depth and clarity to predict PAR at the benthic
surface. Note that aqua_light will still output estimates of PAR at the
stream surface denoted by the column "PAR_surface".
}
