% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azimuth_adj.R
\name{azimuth_adj}
\alias{azimuth_adj}
\title{Helper function for determining the correct solar azimuth}
\usage{
azimuth_adj(driver_file, Lat, Lon)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}
}
\value{
Returns an estimate of azimuth for the current timestep + 1 minute.
}
\description{
This function feeds into the solar_c function and is used to
help determine the correct solar azimuth for locations where latitude is
greater than the solar declination angle. Based on the original solarC.m
matlab code.
}
