% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Function for converting from degrees to radians}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{Angle in degrees}
}
\value{
Returns an angle in radians
}
\description{
Function for converting from degrees to radians
}
