% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_light.R
\name{plot_light}
\alias{plot_light}
\title{Generates rough graphs and daily summaries of model outputs}
\usage{
plot_light(pred_dir, Site)
}
\arguments{
\item{pred_dir}{The read directory where model predictions are located, "C:/}

\item{Site}{The Site_ID}
}
\value{
Returns daily mean (umol m-2 s-1) and total (kmol m-2 d-1) incoming and
predicted light
}
\description{
This function calculates daily mean (umol m-2 s-1) or total (kmol m-2 d-1)
incoming and predicted light
}
