% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_transmission.R
\name{predict_transmission}
\alias{predict_transmission}
\title{Calculate light transmission as a function of depth and and the
irradiance attenuation coefficient}
\usage{
predict_transmission(driver)
}
\arguments{
\item{driver}{The site driver file}
}
\value{
Returns a dataframe of predicted light at the benthic surface
}
\description{
Calculates transmission of light in the water as a function
of depth and the irradiance attenuation coefficient (Kd). The decision of
how to produce a timeseries of Kd is left to the user based on their interests
and available data. An empirical relationship to estimate Kd from turbidity
is presented in the model documentation and associated publication.
}
