% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2deg.R
\name{rad2deg}
\alias{rad2deg}
\title{Function for converting from radians to degrees}
\usage{
rad2deg(x)
}
\arguments{
\item{x}{Angle in radians}
}
\value{
Returns an angle in degrees
}
\description{
Function for converting from radians to degrees
}
