% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_transfer_calc.R
\name{rad_transfer_calc}
\alias{rad_transfer_calc}
\title{Calculate below canopy PAR}
\usage{
rad_transfer_calc(driver, Lat, Lon, x_LAD)
}
\arguments{
\item{driver}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{x_LAD}{Leaf angle distribution, default = 1}
}
\value{
Returns a time series of below canopy PAR
}
\description{
This function calculates below canopy PAR

References
#' \itemize{
  \item Campbell & Norman (1998) An introduction to Environmental biophysics (abbr C&N (1998))
  \item Spitters et al. (1986) Separating the diffuse and direct component of global
        radiation and its implications for modeling canopy photosynthesis: Part I
        components of incoming radiation
  \item Goudriaan (1977)
}
}
