% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHADE2_AL.R
\name{SHADE2_AL}
\alias{SHADE2_AL}
\title{Modified version of the SHADE2 model}
\usage{
SHADE2_AL(
  driver,
  solar,
  Lat,
  Lon,
  channel_azimuth,
  bankfull_width,
  BH,
  TH,
  overhang,
  overhang_height
)
}
\arguments{
\item{driver}{The site driver file}

\item{solar}{Solar geometry, calculated from solar_geo_calc.R}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{channel_azimuth}{Channel azimuth}

\item{bankfull_width}{Bankfull width (m)}

\item{BH}{Bank height (m)}

\item{TH}{Tree height (m)}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}
}
\value{
Returns total percent of the wetted width shaded by the bank and by vegetation
}
\description{
This is a modified version of the SHADE2 function in this package
designed to work with the aqua_light function. The SHADE2 function in this package
is based on the original SHADE2 model from Li et al. (2012) Modeled riparian
stream shading: Agreement with field measurements and sensitivity to riparian conditions
}
