% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade_calc.R
\name{shade_calc}
\alias{shade_calc}
\title{Calculating the percent of the wetted width shaded by banks and vegetation}
\usage{
shade_calc(
  delta,
  solar_altitude,
  bottom_width,
  BH,
  BS,
  WL,
  TH,
  overhang,
  overhang_height
)
}
\arguments{
\item{delta}{difference between the sun and stream azimuth (sun-stream)}

\item{solar_altitude}{Solar altitude (radians) calculated by solar_c.R}

\item{bottom_width}{#ADD DETAILS}

\item{BH}{Bank height}

\item{BS}{Bank slope}

\item{WL}{Water level}

\item{TH}{Tree height}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}
}
\value{
Returns percent of the wetted width shaded by the bank and by vegetation
}
\description{
Translation of r_shade.m
}
