% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shade_calc_AL.R
\name{shade_calc_AL}
\alias{shade_calc_AL}
\title{Calculating the percent of the wetted width shaded by banks and vegetation
modified for use with aqua_light}
\usage{
shade_calc_AL(
  delta,
  solar_altitude,
  water_width,
  WL,
  bankfull_width,
  BH,
  TH,
  overhang,
  overhang_height
)
}
\arguments{
\item{delta}{difference between the sun and stream azimuth (sun-stream)}

\item{solar_altitude}{Solar altitude (radians) calculated by solar_c.R}

\item{water_width}{Wetted water width}

\item{WL}{Water level}

\item{bankfull_width}{Bankfull width}

\item{BH}{Bank height}

\item{TH}{Tree height}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}
}
\value{
Returns percent of the wetted width shaded by the bank and by vegetation
}
\description{
A modified version of the shade_calc function designed to
work with the changes reflected in the aqua_light function.
}
