% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_geo_calc.R
\name{solar_geo_calc}
\alias{solar_geo_calc}
\title{Calculates solar geometry for use in the SHADE2 model}
\usage{
solar_geo_calc(driver_file, Lat, Lon)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}
}
\value{
Returns solar altitude, an initial estimate of azimuth, and declination
}
\description{
This function feeds into the solar_c function and calculates
the solar altitude, solar azimuth, and the solar declination
angle. The calculated solar azimuth is a first estimate that needs to be
adjusted in solar_c based on latitude and the solar declination angle.
Based on the original solarC.m matlab code.
}
