% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_light.R
\name{stream_light}
\alias{stream_light}
\title{Model to predict light at the stream surface}
\usage{
stream_light(
  driver_file,
  Lat,
  Lon,
  channel_azimuth,
  bottom_width,
  BH,
  BS,
  WL,
  TH,
  overhang,
  overhang_height,
  x_LAD
)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{channel_azimuth}{#ADD DETAILS}

\item{bottom_width}{#ADD DETAILS}

\item{BH}{Bank height}

\item{BS}{Bank slope}

\item{WL}{Water level}

\item{TH}{Tree height}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}

\item{x_LAD}{Leaf angle distribution, default = 1}
}
\value{
Returns a time series of predicted light at the stream surface
}
\description{
A combination of the SHADE2 model (Li et al. 2012) and Campbell &
Norman (1998) radiative transfer model
}
