% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_reflection.R
\name{surface_reflection}
\alias{surface_reflection}
\title{Calculate reflection from the water surface}
\usage{
surface_reflection(driver_file, solar_geo)
}
\arguments{
\item{solar_geo}{Solar geometry, calculated from solar_geo_calc.R}

\item{driver}{The site driver file}
}
\value{
Returns the % surface reflection
}
\description{
This function calculates reflection on the water surface based on
solar geometry following Kirk (2011) Light and photosynthesis in
aquatic ecosystems.
}
