% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_CN_1998.R
\name{RT_CN_1998}
\alias{RT_CN_1998}
\title{Calculate below canopy PAR}
\usage{
RT_CN_1998(driver_file, Lat, Lon, x_LAD, ...)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{x_LAD}{Leaf angle distribution, default = 1}
}
\value{
Returns a time series of below canopy PAR
}
\description{
This function calculates below canopy PAR

References
#' \itemize{
  \item Campbell & Norman (1998) An introduction to Environmental biophysics (abbr C&N (1998))
  \item Spitters et al. (1986) Separating the diffuse and direct component of global
        radiation and its implications for modeling canopy photosynthesis: Part I
        components of incoming radiation
  \item Goudriaan (1977)
}
}
