% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHADE2.R
\name{SHADE2}
\alias{SHADE2}
\title{SHADE2 model from Li et al. (2012) Modeled riparian stream shading:
Agreement with field measurements and sensitivity to riparian conditions}
\usage{
SHADE2(
  driver,
  solar,
  Lat,
  Lon,
  channel_azimuth,
  bottom_width,
  BH,
  BS,
  WL,
  TH,
  overhang,
  overhang_height
)
}
\arguments{
\item{driver}{The site driver file}

\item{solar}{Solar geometry, calculated from solar_geo_calc.R}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}

\item{channel_azimuth}{Channel azimuth}

\item{bottom_width}{#ADD DETAILS}

\item{BH}{Bank height}

\item{BS}{Bank slope}

\item{WL}{Water level}

\item{TH}{Tree height}

\item{overhang}{Effectively canopy radius}

\item{overhang_height}{height of the maximum canopy overhang (think canopy radius)}
}
\value{
Returns total percent of the wetted width shaded by the bank and by vegetation
}
\description{
Translation of shdexe.m
}
