% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_c.R
\name{solar_c}
\alias{solar_c}
\title{Calculates solar geometry for use in the SHADE2 model}
\usage{
solar_c(driver_file, solar_geo, Lat, Lon, ...)
}
\arguments{
\item{driver_file}{The model driver file}

\item{solar_geo}{Solar geometry, calculated from solar_geo_calc.R}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}
}
\value{
Returns solar altitude, azimuth, and declination
}
\description{
Translation of solarC.m
}
