% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_geo_calc.R
\name{solar_geo_calc}
\alias{solar_geo_calc}
\title{Calculates solar geometry}
\usage{
solar_geo_calc(driver_file, Lat, Lon)
}
\arguments{
\item{driver_file}{The model driver file}

\item{Lat}{The site Latitude}

\item{Lon}{The site Longitude}
}
\value{
Returns solar declination, altitude, zenith angle, and an initial estimate of azimuth
}
\description{
This function calculates solar declination, altitude,
zenith angle, and an initial estimate of azimuth. This initial estimate
of solar azimuth is passed to the solar_c function where it is adjusted
based on latitude and the solar declination angle. This code is based on
the original solarC.m matlab code.
}
